Reorganize builtin-pages components directory
