+++ app/builtin-pages/com/files-browser/file-editor.js
