+++ app/builtin-pages/com/files-browser/file-preview.js
