+++ app/builtin-pages/com/library/createfile-popup.js
@@ -1,7 +1,7 @@
 /* globals beaker */
 
 import yo from 'yo-yo'
-import closeIcon from '../icon/close'
+import closeIcon from '../../icon/close'
 import {join as joinPaths} from 'path'
 import BINARY_EXTENSIONS from 'binary-extensions'
