+++ app/builtin-pages/com/settings/favicon-picker.js
