+++ app/builtin-pages/views/bookmarks.js
@@ -3,7 +3,7 @@
 const yo = require('yo-yo')
 import {getHostname} from '../../lib/strings'
 import * as toast from '../com/toast'
-import * as editBookmarkPopup from '../com/edit-bookmark-popup'
+import * as editBookmarkPopup from '../com/settings/edit-bookmark-popup'
 import renderBuiltinPagesNav from '../com/builtin-pages-nav'
 import toggleable from '../com/toggleable'
 import renderCloseIcon from '../icon/close'
