+++ app/builtin-pages/views/library-view.js
@@ -7,29 +7,28 @@ import {FSArchive} from 'beaker-virtual-fs'
 import {Archive as LibraryDatArchive} from 'builtin-pages-lib'
 import parseDatURL from 'parse-dat-url'
 import {diffLines} from '@beaker/dat-archive-file-diff'
+import {pluralize, shortenHash} from '@beaker/core/lib/strings'
 import _get from 'lodash.get'
 import throttle from 'lodash.throttle'
 import dragDrop from 'drag-drop'
 import {join as joinPaths} from 'path'
-import FilesBrowser from '../com/files-browser2'
-import toggleable from '../com/toggleable'
+import FilesBrowser from '../com/files-browser/files-browser'
+import {setup as setupAce, isSetup as isAceSetup, config as configureAce, getValue as getAceValue, setValue as setAceValue} from '../com/files-browser/file-editor'
 import toggleable2, {closeAllToggleables} from '../com/toggleable2'
 import renderPeerHistoryGraph from '../com/peer-history-graph'
 import * as contextMenu from '../com/context-menu'
 import * as toast from '../com/toast'
 import * as noticeBanner from '../com/notice-banner'
-import * as localSyncPathPopup from '../com/library-localsyncpath-popup'
-import * as copyDatPopup from '../com/library-copydat-popup'
-import * as createFilePopup from '../com/library-createfile-popup'
-import renderFaviconPicker from '../com/favicon-picker'
+import * as localSyncPathPopup from '../com/library/localsyncpath-popup'
+import * as copyDatPopup from '../com/library/copydat-popup'
+import * as createFilePopup from '../com/library/createfile-popup'
 import renderBackLink from '../com/back-link'
-import renderArchiveHistory from '../com/archive-history'
+import renderArchiveHistory from '../com/archive/archive-history'
 import {RehostSlider} from '../../lib/fg/rehost-slider'
-import LibraryViewCompare from '../com/library-view-compare'
-import LibraryViewLocalCompare from '../com/library-view-local-compare'
-import renderSettingsField from '../com/settings-field'
-import {setup as setupAce, isSetup as isAceSetup, config as configureAce, getValue as getAceValue, setValue as setAceValue} from '../com/file-editor'
-import {pluralize, shortenHash} from '@beaker/core/lib/strings'
+import LibraryViewCompare from '../com/library/view-compare'
+import LibraryViewLocalCompare from '../com/library/view-local-compare'
+import renderSettingsField from '../com/settings/settings-field'
+import renderFaviconPicker from '../com/settings/favicon-picker'
 import {writeToClipboard, findParent} from '../../lib/fg/event-handlers'
 import createMd from '../../lib/fg/markdown'
